requirements.txt
hydra-core
pandas
numpy
matplotlib
stapy
python-Levenshtein
ipython
strenum
frost_sta_client
datetime
geopy
pytest
netCDF4
psycopg2
shapely